add kube logs script

See merge request boxboat/jenkins/jenkins-shared-upstream!26

+++ src/com/boxboat/jenkins/library/LibraryScript.groovy
@@ -6,7 +6,7 @@ class LibraryScript {
 
     static String run(String script) {
         def path = "sharedLibraryScripts/${script}"
-        if (!Config.pipeline.fileExists(path)) {
+        if (Config.pipeline && !Config.pipeline.fileExists(path)) {
             String data = Config.pipeline.libraryResource(resource: "com/boxboat/jenkins/${path}", encoding: "Base64")
             Config.pipeline.writeFile(file: path, text: data, encoding: "Base64")
             Config.pipeline.sh "chmod +x ${path}"
